Embed some properties in the Object cell.
Currently, only a single property is embedded in an object only when it has
no other properties.  Once another property is added, the field for the
first property is used for a pointer to the array for overflow properties.
Only exceptions are objects that are created during initialisation.
They can have multiple embedded properties.

The number of embedded prpeprties can be specified each time when an
individual simple objects are created. For other kinds of objects,
the number of embedded properties are fixed to 1 because other fields
follows to the embedded property field. The number cannot be changed
dynamically, which may require read barriers to redirect.