Put __proto__ in hidden classes for objects created by user-defined ctors.
User-defined constructor function has its own "prototype".
The hidden class object for objects created by the constructor has
this prototype object in its __proto__ field.  To search for this
hidden class object, the prototype object has a refrence to the
hidden class in "__hidden_class__" property, which is intended to
be a hidden (but is implemented as a normal) property.
This prototype object - hidden class pair becomes garbage together.