New instruction creates pre-formed objects if statistics are available.
We introduce variations of hidden class: base and preformed.
The base hidden classes are on the hidden class trnasition graph
starting from the root. The preformed one is derived from a
base one to embed more properties in objects.
A base one is always created before a preformed one associated
with it is created. AllocSite has both base one (`hc') and
preformed one (`preformed_hc'). The base one is used to find
LUB of possible hidden classes on the hidden class transistion
graph, which preformed one is used to initialise the object
being created.
Note that preformed one is created in the New instruction
on-demand because we cannot create it in GC.