Stop applying no black to while invariant check to HTAG_ARRAY_DATA.
Array data may have uninitialised part. Only indexes from 0
to array.length are guaranteed to be initliased. However,
GC cannot know the value of array.length while cheking this
invariant because GC only has a pointer to the array data,
which does not has a pointer to the array header.

If uninitialised part has pointers to white objects, this
invariant cheker detects a false error. The simplest solution
is to ignore array data objects.