Allocate HashTable and HashCell in GC heap.
In the previous version, HashTable and HashCell are allocated
using malloc. GC did trace them to find their children, but
it did not mark them because they are GC heap objects. And
they are not freed either automatically or manually.
In addition, GC did not care about the cases where these objects
are registered as GC roots using GC_PUSH.